+++ src/pricecypher/__init__.py
@@ -1,12 +1,10 @@
 import pbr.version
 
-from .collections import *
-from .collections.scope_collection import ScopeCollection
-from .collections.scope_value_collection import ScopeValueCollection
+from .collections import ScopeCollection, ScopeValueCollection
 from .config_sections import ConfigSections
-from .contracts import *
 from .datasets import Datasets
-from .oidc import *
-from .rest import RestClientOptions, RestClient
+from .rest import RestClient
+
+__all__ = [ScopeCollection, ScopeValueCollection, ConfigSections, Datasets, RestClient]
 
 __version__ = pbr.version.VersionInfo('pricecypher_sdk').version_string()
