+++ src/pricecypher/contracts/dataclasses/__init__.py
@@ -1,2 +1,4 @@
-from .PredictResult import *
-from .TestResult import *
+from .predict_result import PredictStep, PredictValues, PredictResult
+from .test_result import ElementTestResult, ElementTest, TestSuite, TestResult
+
+__all__ = [PredictStep, PredictValues, PredictResult, ElementTestResult, ElementTest, TestSuite, TestResult]
