+++ src/pricecypher/contracts/dataclasses/predict_result.py
