--- src/pricecypher/contracts/enums/__init__.py
@@ -1,2 +1,0 @@
-from .Accuracy import Accuracy
-from .TestStatus import TestStatus
