+++ src/pricecypher/endpoints/__init__.py
@@ -1,3 +1,5 @@
 from .config import ConfigEndpoint
 from .datasets import DatasetsEndpoint
 from .users import UsersEndpoint
+
+__all__ = [DatasetsEndpoint, UsersEndpoint, ConfigEndpoint]
