+++ src/pricecypher/enums/accuracy.py
