+++ src/pricecypher/models/__init__.py
@@ -1,3 +1,7 @@
-from .config import *
-from .datasets import *
-from .users import *
+from .config import ConfigSection, ConfigSectionWithKeys
+from .datasets import Scope, ScopeValue, TransactionSummary, ScopeValueTransaction, ScopeConstantTransaction, \
+    Transaction, PageMeta, TransactionsPage
+from .users import Dataset
+
+__all__ = [Dataset, Scope, ScopeValue, TransactionSummary, ScopeValueTransaction, ScopeConstantTransaction, Transaction,
+           PageMeta, TransactionsPage, ConfigSection, ConfigSectionWithKeys]
